refactor: move nags out of main project member header for perf
