+++ apps/frontend/src/locales/en-US/index.json
@@ -557,9 +557,6 @@
   "project-member-header.accept": {
     "message": "Accept"
   },
-  "project-member-header.action": {
-    "message": "Action"
-  },
   "project-member-header.decline": {
     "message": "Decline"
   },
@@ -581,33 +578,6 @@
   "project-member-header.invitation-with-role": {
     "message": "You've been invited be a member of this project with the role of '{role}'."
   },
-  "project-member-header.key-title": {
-    "message": "Status Key"
-  },
-  "project-member-header.publishing-checklist": {
-    "message": "Publishing checklist"
-  },
-  "project-member-header.required": {
-    "message": "Required"
-  },
-  "project-member-header.resubmit-for-review": {
-    "message": "Resubmit for review"
-  },
-  "project-member-header.resubmit-for-review-desc": {
-    "message": "Your project has been {status} by Modrinth's staff. In most cases, you can resubmit for review after addressing the staff's message."
-  },
-  "project-member-header.show-key": {
-    "message": "Toggle key"
-  },
-  "project-member-header.submit-checklist-tooltip": {
-    "message": "You must complete the required steps in the publishing checklist!"
-  },
-  "project-member-header.submit-for-review": {
-    "message": "Submit for review"
-  },
-  "project-member-header.submit-for-review-desc": {
-    "message": "Your project is only viewable by members of the project. It must be reviewed by moderators in order to be published."
-  },
   "project-member-header.success": {
     "message": "Success"
   },
@@ -617,13 +587,34 @@
   "project-member-header.success-join": {
     "message": "You have joined the project team"
   },
-  "project-member-header.suggestion": {
+  "project-moderation-nags.publishing-checklist": {
+    "message": "Publishing checklist"
+  },
+  "project-moderation-nags.required": {
+    "message": "Required"
+  },
+  "project-moderation-nags.resubmit-for-review": {
+    "message": "Resubmit for review"
+  },
+  "project-moderation-nags.resubmit-for-review-desc": {
+    "message": "Your project has been {status} by Modrinth's staff. In most cases, you can resubmit for review after addressing the staff's message."
+  },
+  "project-moderation-nags.submit-checklist-tooltip": {
+    "message": "You must complete the required steps in the publishing checklist!"
+  },
+  "project-moderation-nags.submit-for-review": {
+    "message": "Submit for review"
+  },
+  "project-moderation-nags.submit-for-review-desc": {
+    "message": "Your project is only viewable by members of the project. It must be reviewed by moderators in order to be published."
+  },
+  "project-moderation-nags.suggestion": {
     "message": "Suggestion"
   },
-  "project-member-header.visit-moderation-page": {
+  "project-moderation-nags.visit-moderation-page": {
     "message": "Visit moderation page"
   },
-  "project-member-header.warning": {
+  "project-moderation-nags.warning": {
     "message": "Warning"
   },
   "project-type.collection.plural": {
